Get Version Information on Files of Given IDs

Get Version Information on Files of Given IDs

Usage

get_version_info(
  file_ids,
  user = nextcloud_user(),
  ignore = "^(status|quota|getcontent|resource)",
  dbg = TRUE
)

Arguments

file_ids

vector of integer representing file ids (values returned in column fileid) by list_files

user

nextcloud user

ignore

pattern matching columns to be excluded from the result. Set to NULL to see what columns are available.

dbg

logical indicating whether or not to print debug messages

Value

data frame with one row per version. There seems to be only an entry if the corresponding file as more than one version.